+++ jordan/lib/views/plugins/pluginPrayer_screen.dart
@@ -7,7 +7,7 @@ import 'package:jordan/views/widgets/htmlBuilder_widget.dart';
 
 import 'dart:async' show Future;
 
-import 'package:jordan/views/widgets/taskView_widget.dart';
+import 'package:jordan/views/widgets/tabView_widget.dart';
 
 Future<String> loadAsset(BuildContext context, String asset) async {
   return await DefaultAssetBundle.of(context).loadString(asset);
